add Filename#role_to_abbreviation
Providing a means to access the "coerced role" string. Here's the
situation: there are times, e.g., for Bookeye4 books that only have
dmaker files, when one wants to use a file in multiple roles, e.g.,
dmaker as dmaker AND master. Previously, the "role" was strictly
determined by the filename. I just added the ability to override the
role, and now I need a way to access the letter representation for the
role so that I can generate unique METS identifiers for the files.
This commit adds the Filename#role_to_abbreviation method that
allows legacy code to work as before, and new code to be written.